# doc/openapi3.yaml
openapi: 3.0.4
info:
  title: Orticello API
  version: 1.0.0
  description: API for managing community gardens (orti urbani) in Trento
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://orticelloapi.onrender.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AffidaLotto:
      type: object
      required:
        - lotto
        - utente
        - data_inizio
        - data_fine
      properties:
        _id:
          type: string
          description: MongoDB ObjectId (generated automatically)
          readOnly: true
          example: 507f1f77bcf86cd799439017
        lotto:
          type: string
          description: Reference to Lotto ObjectId
          example: 507f1f77bcf86cd799439012
        utente:
          type: string
          description: Reference to Utente ObjectId
          example: 507f1f77bcf86cd799439016
        data_inizio:
          type: string
          format: date
          description: Assignment start date
          example: 2025-01-01
        data_fine:
          type: string
          format: date
          description: Assignment end date
          example: 2025-12-31
        colture:
          type: array
          items:
            type: string
          description: List of crops planted by the user
          example:
            - Pomodori
            - Zucchine
            - Insalata
        createdAt:
          type: string
          format: date-time
          description: Timestamp of creation (generated automatically)
          readOnly: true
          example: 2025-12-11T10:30:00.000Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last update (generated automatically)
          readOnly: true
          example: 2025-12-11T10:30:00.000Z
...
paths:
  /api/v1/affidaLotti:
    get:
      summary: Get list of lotto assignments
      description: Retrieve a list of all lotto assignments (affidaLotti). Returns an
        array showing which users have been assigned to which lotti, including
        assignment dates and optional colture information. Accessible to
        **Comune** and **Associazioni**.
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully retrieved list of affida lotti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AffidaLotto"
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Error retrieving affida lotti
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error retrieving affida lotti
                  error:
                    type: object
    post:
      summary: Create a new lotto assignment
      description: Create a new assignment of a lotto to a user with start and end
        dates, and optionally a list of crops. Accessible to authenticated users
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffidaLotto"
      responses:
        "201":
          description: AffidaLotto created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffidaLotto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid input data
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Error creating affida lotto
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error creating affida lotto
                  error:
                    type: object
  /api/v1/affidaLotti/attivi:
    get:
      summary: Get active lotto assignments
      description: Retrieve only lotto assignments that have already started and not
        yet ended (data_inizio <= now <= data_fine). Results include populated
        lotto and utente data and are sorted by end date. Accessible to
        **Comune** and **Associazini**
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully retrieved list of active affida lotti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AffidaLotto"
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Error retrieving active affida lotti
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error retrieving active affida lotti
                  error:
                    type: object
  /api/v1/affidaLotti/{id}:
    get:
      summary: Get lotto assignment by ID
      description: Retrieve a single lotto assignment by its unique identifier
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the affida lotto
      responses:
        "200":
          description: Successfully retrieved affida lotto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffidaLotto"
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "404":
          description: AffidaLotto not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto not found
        "500":
          description: Error retrieving affida lotto
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error retrieving affida lotto
                  error:
                    type: object
    put:
      summary: Update lotto assignment by ID
      description: Update an existing lotto assignment with new information (dates,
        colture, etc.). Used by **Comune** and **Associazioni**
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the affida lotto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffidaLotto"
      responses:
        "200":
          description: AffidaLotto updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffidaLotto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid input data
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "404":
          description: AffidaLotto not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto not found
        "500":
          description: Error updating affida lotto
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error updating affida lotto
                  error:
                    type: object
    delete:
      summary: Delete lotto assignment by ID
      description: Remove an existing lotto assignment from the system. Accessible ony
        to **Comune** and **Associazioni**
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the affida lotto
      responses:
        "200":
          description: AffidaLotto deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto deleted successfully
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "404":
          description: AffidaLotto not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto not found
        "500":
          description: Error deleting affida lotto
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error deleting affida lotto
                  error:
                    type: object
  /api/v1/affidaLotti/{id}/colture:
    post:
      summary: Add a crop to lotto assignment
      description: Add a new crop (coltura) to the list of crops for a specific lotto
        assignment. Accessible only to the **Cittadino** owning the assignment
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the affida lotto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coltura:
                  type: string
                  description: Name of the crop to add
                  example: Pomodori
      responses:
        "200":
          description: Coltura added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffidaLotto"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Coltura name is required
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "404":
          description: AffidaLotto not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto not found
        "500":
          description: Error adding coltura
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error adding coltura
                  error:
                    type: object
  /api/v1/affidaLotti/{id}/colture/{coltura}:
    delete:
      summary: Remove a crop from lotto assignment
      description: Remove a specific crop (coltura) from the list of crops for a lotto
        assignment
      tags:
        - AffidaLotti
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the affida lotto
        - in: path
          name: coltura
          required: true
          schema:
            type: string
          description: Name of the crop to remove
      responses:
        "200":
          description: Coltura removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffidaLotto"
        "401":
          description: Unauthorized - Authentication required (missing or invalid JWT token)
        "403":
          description: Forbidden - Insufficient permissions
        "404":
          description: AffidaLotto not found or coltura not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AffidaLotto not found or coltura not found
        "500":
          description: Error removing coltura
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error removing coltura
                  error:
                    type: object
...